+++ src/run/audit.rs
@@ -5,8 +5,8 @@ use std::fs;
 use std::io::Read;
 use std::path::{Path, PathBuf};
 
-use crate::cli::args::AuditCmd;
 use crate::error::Error;
+use crate::run::args::AuditCmd;
 
 /// Recompute sweep-wide aggregates and reconcile with results.json and evaluation.json.
 #[allow(clippy::too_many_lines)]
